fix display of button when toggle
