+++ src/component/Collect.tsx
@@ -1,7 +1,8 @@
-import { Dispatch, SetStateAction } from "react";
 import styled from "styled-components";
 import { db } from "../utils/firebase";
 import { doc, updateDoc, getDoc, DocumentData } from "firebase/firestore";
+import { useContext, useEffect } from "react";
+import { GlobalContext } from "../App";
 
 const CollectButton = styled.div`
   width: 64px;
@@ -29,15 +30,9 @@ const CollectedButton = styled(CollectButton)`
   }
 `;
 
-const Collect = ({
-  postId,
-  isSaved,
-  setIsSaved,
-}: {
-  postId: string;
-  isSaved: boolean;
-  setIsSaved?: Dispatch<SetStateAction<boolean>>;
-}) => {
+const Collect = ({ postId }: { postId: string }) => {
+  const st: any = useContext(GlobalContext);
+
   // const userId = take users doc.id (which user will get when successfully sign up or sign in, doc.id also equals to user.id)
   const modifyCollect = async () => {
     const docRef = doc(db, "/users/RuJg8C2CyHSbGMUwxrMr");
@@ -45,20 +40,34 @@ const Collect = ({
     let rawUserCollection = docSnap.data().user_collection;
     let updateUserPost;
 
-    if (isSaved) {
+    if (st.isSaved) {
       updateUserPost = rawUserCollection.filter(
         (item: string) => item !== postId
       );
       await updateDoc(docRef, { user_collection: updateUserPost });
-      setIsSaved && setIsSaved(false);
+      st.setIsSaved(false);
     } else {
       updateUserPost = [...rawUserCollection, postId];
       await updateDoc(docRef, { user_collection: updateUserPost });
-      setIsSaved && setIsSaved(true);
+      st.setIsSaved(true);
     }
   };
 
-  return isSaved ? (
+  useEffect(() => {
+    const getData = async () => {
+      const docRef = doc(db, "users/RuJg8C2CyHSbGMUwxrMr");
+      const docSnap: DocumentData = await getDoc(docRef);
+      const userCollection = docSnap.data().user_collection;
+      if (userCollection.includes(postId)) {
+        st.setIsSaved(true);
+      } else {
+        st.setIsSaved(false);
+      }
+    };
+    getData();
+  }, []);
+
+  return st.isSaved ? (
     <CollectedButton
       onClick={(e) => {
         e.stopPropagation();
